HTBHF-413 move dateAsString() to formatters
