+++ src/web/routes/application/common/formatters.js
@@ -9,6 +9,19 @@ const toDateString = (day, month, year) => {
   return [year, parseMonth, parseDay].join('-')
 }
 
+const dateAsString = ({ date = new Date(), monthAdjustment = 0 } = {}) => {
+  if (typeof monthAdjustment !== 'number') {
+    throw new Error('Month adjustment must be numeric')
+  }
+  const dateToChange = new Date(date)
+  dateToChange.setMonth(dateToChange.getMonth() + monthAdjustment)
+  const dd = dateToChange.getDate()
+  const mm = dateToChange.getMonth() + 1
+  const yyyy = dateToChange.getFullYear()
+  return toDateString(dd, mm, yyyy)
+}
+
 module.exports = {
-  toDateString
+  toDateString,
+  dateAsString
 }
