+++ src/web/routes/application/common/validators.js
@@ -1,23 +1,11 @@
 const { isNil } = require('ramda')
 const validator = require('validator')
-const { toDateString } = require('./formatters')
+const { dateAsString } = require('./formatters')
 
 const DATE_PATTERN = /^[0-9]{4}-[0-9]{2}-[0-9]{2}$/
 
 const isValidDate = (dateString) =>
   !isNil(dateString) && validator.matches(dateString, DATE_PATTERN) && validator.isISO8601(dateString, { strict: true })
-
-const dateAsString = ({ date = new Date(), monthAdjustment = 0 } = {}) => {
-  if (typeof monthAdjustment !== 'number') {
-    throw new Error('Month adjustment must be numeric')
-  }
-  const dateToChange = new Date(date)
-  dateToChange.setMonth(dateToChange.getMonth() + monthAdjustment)
-  const dd = dateToChange.getDate()
-  const mm = dateToChange.getMonth() + 1
-  const yyyy = dateToChange.getFullYear()
-  return toDateString(dd, mm, yyyy)
-}
 
 const isDateInPast = (dateString) => {
   if (isValidDate(dateString)) {
@@ -44,6 +32,5 @@ module.exports = {
   isValidDate,
   isDateInPast,
   isDateMoreThanOneMonthAgo,
-  isDateMoreThanEightMonthsInTheFuture,
-  dateAsString
+  isDateMoreThanEightMonthsInTheFuture
 }
